feat: renames app
