+++ .circleci/config.yml
@@ -80,8 +80,8 @@ jobs:
           name: Deploy to AWS S3
           command: |
             aws lambda update-function-code \
-              --function-name=banshee-44-mods-bot \
-              --zip-file=fileb://build/banshee-44-mods-bot.zip 1> /dev/null \
+              --function-name=destiny-insights-bot \
+              --zip-file=fileb://build/destiny-insights-bot.zip 1> /dev/null \
               --region=us-east-1
 
 workflows:
