+++ README.md
@@ -1,29 +1,38 @@
 ![logo](./media/github-readme-banner.png)
 
 <p align="center">
-  <a href="https://circleci.com/gh/cujarrett/banshee-44-mods-bot/tree/main"><img alt="Circle CI" src="https://circleci.com/gh/cujarrett/banshee-44-mods-bot/tree/main.svg?style=svg"></a>
-  <a href="https://twitter.com/intent/follow?screen_name=banshee44mods"><img alt="Follow on Twitter" src="https://img.shields.io/twitter/follow/banshee44mods?label=Follow&style=social"></a>
+  <a href="https://circleci.com/gh/cujarrett/destiny-insights-bot/tree/main"><img alt="Circle CI" src="https://circleci.com/gh/cujarrett/destiny-insights-bot/tree/main.svg?style=svg"></a>
+  <a href="https://twitter.com/intent/follow?screen_name=destinyinsights"><img alt="Follow on Twitter" src="https://img.shields.io/twitter/follow/destinyinsights?label=Follow&style=social"></a>
   <a href="https://discord.gg/jAA5U52"><img alt="Chat on Discord" src="https://img.shields.io/discord/460598989939802115?label=Discord"></a>
   <a href="https://github.com/semantic-release/semantic-release"><img alt="Project uses semantic-release" src="https://img.shields.io/badge/%20%20%F0%9F%93%A6%F0%9F%9A%80-semantic--release-e10079.svg"></a>
   <a href="http://commitizen.github.io/cz-cli/"><img alt="Commitizen friendly" src="https://img.shields.io/badge/commitizen-friendly-brightgreen.svg?"></a>
 </p>
 
 ## What is it?
-It's a serverless Twitter bot powered by JavaScript and AWS that tweets Destiny 2's Banshee-44 daily mods available for sale. Check
-out [Banshee-44's current mods for sale right now on Twitter](https://twitter.com/banshee44mods).
+
+It's a serverless Twitter bot powered by JavaScript and AWS that tweets Destiny
+2's note worthy items such as mods, high stat armor, or well rolled weapons
+available for sale. Check out
+[@destinyinsights right now on Twitter](https://twitter.com/destinyinsights).
 
 ## Questions & Contact
-Feel free to join the [Banshee-44 Mods Bot Discord](https://discord.gg/jAA5U52) if you have questions.
+
+Feel free to join the [Discord](https://discord.gg/jAA5U52) if you have
+questions.
 
 ## Want to Contribute?
+
 [Check out the contributing guide](CONTRIBUTING.md)
 
 ## Thank You!
-- Big thanks to [Tom Chapman](https://github.com/justrealmilk) for his help with Bungie's API.
+
+Big thanks to [Tom Chapman](https://github.com/justrealmilk) for his help with
+Bungie's API.
 
 ## Changelog
-[banshee-44-mods-bot changelog](CHANGELOG.md)
+
+[destiny-insights-bot changelog](CHANGELOG.md)
 
 <p align="center">
-  Made with :heart:, JavaScript, and GitHub. Powered by <a href="https://braytech.org/">Braytech</a> and AWS.
+  Made with :heart:, JavaScript, and GitHub. Powered by AWS.
 </p>
