+++ package.json
@@ -1,17 +1,17 @@
 {
-  "name": "banshee-44-mods-bot",
+  "name": "destiny-insights-bot",
   "version": "1.33.0",
   "author": "Matt Jarrett",
   "license": "MIT",
-  "description": "Node JS Twitter bot that tweets Destiny 2's Banshee-44 daily mods available for sale",
+  "description": "Node JS Twitter bot that tweets Destiny's mods, armor, and weapons available",
   "main": "src/index.js",
   "scripts": {
     "commit": "git-cz",
     "lint": "eslint .",
     "fix-lint": "eslint . --fix",
     "test": "nyc --reporter=lcov tape test/**/*.js | tap-spec && nyc report",
-    "build": "npm install --production && rm -rf build && mkdir build && zip -r -q -x='*media*' -x='package-lock.json' -x='*terraform*' -x=*coverage* -x='*.md' -x='LICENSE' -x='*build*' -x='*test*' -x='*.DS_Store*' -x='*.nyc_output*' -x='*.git*' -x='*.nycrc' -x='*.eslintrc.js' -x='release.config.js' -x='commitlint.config.js' build/banshee-44-mods-bot.zip . && du -sh build",
-    "deploy": "aws lambda update-function-code --function-name=banshee-44-mods-bot --zip-file=fileb://build/banshee-44-mods-bot.zip --region=us-east-1 1> /dev/null",
+    "build": "npm install --production && rm -rf build && mkdir build && zip -r -q -x='*media*' -x='package-lock.json' -x='*terraform*' -x=*coverage* -x='*.md' -x='LICENSE' -x='*build*' -x='*test*' -x='*.DS_Store*' -x='*.nyc_output*' -x='*.git*' -x='*.nycrc' -x='*.eslintrc.js' -x='release.config.js' -x='commitlint.config.js' build/destiny-insights-bot.zip . && du -sh build",
+    "deploy": "aws lambda update-function-code --function-name=destiny-insights-bot --zip-file=fileb://build/destiny-insights-bot.zip --region=us-east-1 1> /dev/null",
     "ci": "npm run lint && npm run test",
     "cd": "npm run build && npm run deploy",
     "semantic-release": "semantic-release"
@@ -24,29 +24,29 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/cujarrett/banshee-44-mods-bot.git"
+    "url": "https://github.com/cujarrett/destiny-insights-bot.git"
   },
   "keywords": [
     "Destiny"
   ],
   "bugs": {
-    "url": "https://github.com/cujarrett/banshee-44-mods-bot/issues"
+    "url": "https://github.com/cujarrett/destiny-insights-bot/issues"
   },
-  "homepage": "https://github.com/cujarrett/banshee-44-mods-bot#readme",
+  "homepage": "https://github.com/cujarrett/destiny-insights-bot#readme",
   "devDependencies": {
-    "@commitlint/cli": "~12.0.1",
-    "@commitlint/config-conventional": "~12.0.1",
+    "@commitlint/cli": "~12.1.1",
+    "@commitlint/config-conventional": "~12.1.1",
     "@semantic-release/changelog": "~5.0.1",
     "@semantic-release/commit-analyzer": "~8.0.1",
     "@semantic-release/exec": "~5.0.0",
     "@semantic-release/git": "~9.0.0",
     "@semantic-release/github": "~7.2.0",
-    "@semantic-release/npm": "~7.0.10",
+    "@semantic-release/npm": "~7.1.0",
     "@semantic-release/release-notes-generator": "~9.0.2",
-    "aws-sdk": "~2.863.0",
+    "aws-sdk": "~2.879.0",
     "commitizen": "~4.2.3",
     "cz-conventional-changelog": "~3.3.0",
-    "eslint": "~7.22.0",
+    "eslint": "~7.23.0",
     "husky": "~4.3.8",
     "nconf": "~0.11.2",
     "nyc": "~15.1.0",
