+++ terraform/modules/iam/variables.tf
@@ -1,9 +1,9 @@
-variable "error-sns-topic" {
+variable "error_sns_topic" {
   description = "SNS Topic ARN to trigger on lambda failure"
   type = string
 }
 
-variable "parameter-store-twitter-auth-arn" {
+variable "parameter_store_twitter_auth_arn" {
   description = "Parameter Store twitter auth config path ARN"
   type = string
 }
