+++ terraform/provider.tf
@@ -1,4 +1,3 @@
 provider "aws" {
-  version = "~> 3.0"
   region = var.aws_region
 }
