+++ terraform/variables.tf
@@ -4,12 +4,12 @@ variable "aws_region" {
   default = "us-east-1"
 }
 
-variable "error-sns-topic" {
+variable "error_sns_topic" {
   description = "SNS Topic ARN to trigger on lambda failure"
   type = string
 }
 
-variable "parameter-store-twitter-auth-arn" {
+variable "parameter_store_twitter_auth_arn" {
   description = "Parameter Store twitter auth config path ARN"
   type = string
 }
